compiles -- fixed types

+++ runtime/devnet/src/lib.rs
@@ -731,10 +731,10 @@ impl pallet_motion::Config for Runtime {
 }
 
 parameter_types! {
-	pub const MaxString: u8 = 100;
-	pub const MaxHash: u8 = 512;
-	pub const MaxCredentialsTypes: u8 = 50;
-	pub const MaxServices: u8 = 10;
+	pub const MaxString: u32 = 100;
+	pub const MaxHash: u32 = 512;
+	pub const MaxCredentialsTypes: u32 = 50;
+	pub const MaxServices: u32 = 10;
 	pub const DidDeposit: Balance = 10 * WATRD;
 }
